[refactor] onNewIntent 콜백 -> Consumer 를 사용하는 방법으로 migration
좀 더 Compose 스러운
