+++ feature/main/src/main/AndroidManifest.xml
@@ -5,7 +5,8 @@
 
         <activity
             android:name=".MainActivity"
-            android:exported="false"
+            android:exported="true"
+            android:launchMode="singleTop"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.Unifest">
