+++ feature/main/src/main/kotlin/com/unifest/android/feature/main/MainScreen.kt
@@ -1,5 +1,7 @@
 package com.unifest.android.feature.main
 
+import android.content.Intent
+import androidx.activity.ComponentActivity
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,15 +9,13 @@ import androidx.compose.material3.SnackbarDuration
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
+import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color.Companion.White
 import androidx.compose.ui.platform.LocalContext
-import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.core.util.Consumer
 import androidx.navigation.compose.NavHost
 import com.unifest.android.core.common.UiText
 import com.unifest.android.core.designsystem.component.UnifestScaffold
@@ -31,19 +31,17 @@ import com.unifest.android.feature.waiting.navigation.waitingNavGraph
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
+import timber.log.Timber
 
 private const val SnackBarDuration = 1000L
 
 @Composable
 internal fun MainScreen(
     navigator: MainNavController = rememberMainNavController(),
-    viewModel: MainViewModel = hiltViewModel(),
 ) {
     val snackBarState = remember { SnackbarHostState() }
     val scope = rememberCoroutineScope()
     val context = LocalContext.current
-    val waitingId by viewModel.waitingId.collectAsState()
-    val boothId by viewModel.boothId.collectAsState()
 
     val onShowSnackBar: (message: UiText) -> Unit = { message ->
         scope.launch {
@@ -58,19 +56,7 @@ internal fun MainScreen(
         }
     }
 
-    LaunchedEffect(waitingId) {
-        if (waitingId != 0L) {
-            navigator.navigate(MainTab.WAITING)
-            viewModel.setWaitingId(0L)
-        }
-    }
-
-    LaunchedEffect(boothId) {
-        if (boothId != 0L) {
-            navigator.navigateToBoothDetail(viewModel.boothId.value)
-            viewModel.setBoothId(0L)
-        }
-    }
+    HandleNewIntent(navigator = navigator)
 
     UnifestScaffold(
         bottomBar = {
@@ -140,5 +126,33 @@ internal fun MainScreen(
                 onShowSnackBar = onShowSnackBar,
             )
         }
+    }
+}
+
+@Composable
+private fun HandleNewIntent(navigator: MainNavController) {
+    val activity = LocalContext.current as ComponentActivity
+    DisposableEffect(Unit) {
+        val onNewIntentConsumer = Consumer<Intent> { intent ->
+            Timber.d("onNewIntent -> $intent")
+            Timber.d("onNewIntent -> waitingId: ${intent.getStringExtra("waitingId")}")
+            Timber.d("onNewIntent -> boothId: ${intent.getStringExtra("boothId")}")
+
+            if (intent.getStringExtra("waitingId") != null) {
+                val waitingId = intent.getStringExtra("waitingId")
+                Timber.d("navigate_to_waiting -> waitingId: $waitingId")
+                if (waitingId != null) {
+                    navigator.navigate(MainTab.WAITING)
+                }
+            } else if (intent.getStringExtra("boothId") != null) {
+                val boothId = intent.getStringExtra("boothId")
+                Timber.d("navigate_to_booth -> boothId: $boothId")
+                if (boothId != null) {
+                    navigator.navigateToBoothDetail(boothId.toLong())
+                }
+            }
+        }
+        activity.addOnNewIntentListener(onNewIntentConsumer)
+        onDispose { activity.removeOnNewIntentListener(onNewIntentConsumer) }
     }
 }
