+++ feature/splash/src/main/kotlin/com/unifest/android/feature/splash/SplashActivity.kt
@@ -1,6 +1,7 @@
 package com.unifest.android.feature.splash
 
 import android.annotation.SuppressLint
+import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
